requirements.txt
Flask==2.3.3
Pillow==9.2.0
requests==2.31.0
samsungtvws==2.6.0
